Use pixels for base font size, ems for elements.
The existing sizes are in a confusing mix of points and percentages
which will appear inconsistent across browser (e.g. if someone is not
using the default font size, the percentage based font sizes are all
different).

Now the font sizes should be consistent - especially wrt. to code which
varies across browsers. It's also a lot easier to increase the size of
the text, since only the base size needs to change.

There are possibly problems if the font sizes nest, e.g. for nested
tables.