Added reference.xml auto-generation from dtdxml
This should have been an established process, but for some reason it was
edited manually after initial generation.

After the change `reference.xml` can be removed from the repository, but
I left it to show that the things are working fine and the difference from
an auto-generated one are simple divergences due to manual changing
(not updated `reference.xml` with added `<purpose>` elements in 38017203f69cec3aa5f402fe8306eb1deff67b9a). Also removed the date numbers
from `dtdxml` as they were never in `xml` and seems to have no actual use,
but brings a lot of noise to `xml` differences after added auto-generation.