Fix duplicate equal signs in Doxygen-generated enum value initializers.
Doxygen 1.9.1 generates enum <initializer> tags with text starting with
an equal sign ('=') followed by the enum value initializer. This sign
was copied under the BoostBook <default> tag, and BoostBook styleseets
add an equal sign of their own when converting this tag to HTML. This
resulted in duplicate equal signs in the HTML output.

This commit strips the leading equal sign from the initializer before
putting it into the <default> tag.